Fix IMAP UTF-7 for code points >= U+10000
The 20-year old utf7 conversion functions punted on those values,
which was understandable for when they were written.  We now have
emojis and increasing number of characters that might be used in a
mailbox.

Add encode/decode logic using UTF-16 surrogate pairs.

Thanks to Jeff Sipek for reporting this issue, and for giving pointers
about how the values should be handled.

Upstream-commit: https://gitlab.com/muttmua/mutt/commit/e832240e8e9ccabdcd067553d6f4bf4f567dfaec

Co-authored-by: Kevin McCarthy <kevin@8t8.us>