Fix compare_uid() to work with large UID values
The function was pulled from the other sort methods used in neomutt.  But
those don't work properly for a 32-bit unsigned value.  If the
difference between two UID values is greater than a signed int can
represent, it will sort improperly.

Some of the other sort functions need to be fixed too, but that can be
done in master.

Upstream-commit: https://gitlab.com/muttmua/mutt/commit/a6d83ff61c1f9c26fa504e27c8fefbc6f675a6b9

Co-authored-by: Kevin McCarthy <kevin@8t8.us>